Try all addresses for remote connection
This commit adds a loop over all the found addresses from the user
provided server name. It is possible that we can't connect to one of
the returned connection types so this adds a for loop over all the
results and the first one which establishes a connection is used.
If no connection can be established a ConfigurationError is raised.