Remove abspath() call on play file function
This commit fixes a bug in the player_play_file() function by removing
a call to abspath() before passing it to cmus. This breaks support
for remote files and makes assumptions about what's being passed in.
Instead of that this just assumes that the user knows what they're
passing into cmus.